Fix UB (signed overflow) in QSvgIconEnginePrivate::pmcKey()
Found by UBsan:

  qsvgiconengine.cpp:64:84: runtime error: left shift of 537133056 by 4 places cannot be represented in type 'int'

Fix by casting to qint64 at the start of the shift-or chain.
The total cumulative shift is by 28 places, so 64 bits should
be enough to not overflow anymore.

+++ src/plugins/iconengines/svgiconengine/qsvgiconengine.cpp
@@ -61,7 +61,7 @@ public:
     QString pmcKey(const QSize &size, QIcon::Mode mode, QIcon::State state)
         { return QLatin1String("$qt_svgicon_")
                  + QString::number(serialNum, 16).append(QLatin1Char('_'))
-                 + QString::number((((((size.width()<<11)|size.height())<<11)|mode)<<4)|state, 16); }
+                 + QString::number((((((qint64(size.width()) << 11) | size.height()) << 11) | mode) << 4) | state, 16); }
 
     void stepSerialNum()
         { serialNum = lastSerialNum.fetchAndAddRelaxed(1); }
